+++ src/fs/file.ts
@@ -1,5 +1,5 @@
 // TODO more
 
-export interface FileFilter {
+export interface File_Filter {
 	(id: string): boolean;
 }