+++ src/fs/node.test.ts
@@ -7,17 +7,17 @@ import {fs as node_fs} from './node.js';
 const test_find_files = suite('find_files', {fs: node_fs});
 
 test_find_files('basic behavior', async ({fs}) => {
-	const ignoredPath = 'test1.foo.ts';
-	let hasIgnoredPath = false;
+	const ignored_path = 'test1.foo.ts';
+	let has_ignored_path = false;
 	const result = await fs.find_files(
 		'./src/fs/fixtures',
 		({path}) => {
-			if (!hasIgnoredPath) hasIgnoredPath = path === ignoredPath;
-			return path !== ignoredPath;
+			if (!has_ignored_path) has_ignored_path = path === ignored_path;
+			return path !== ignored_path;
 		},
 		(a, b) => -a[0].localeCompare(b[0]),
 	);
-	t.ok(hasIgnoredPath); // makes sure the test isn't wrong
+	t.ok(has_ignored_path); // makes sure the test isn't wrong
 	t.equal(Array.from(result.keys()), [
 		'test2.foo.ts',
 		'baz2/test2.baz.ts',