+++ src/fs/watch_node_fs.ts
@@ -1,11 +1,11 @@
-import CheapWatch from 'cheap-watch';
+import Cheap_Watch from 'cheap-watch';
 import {omit_undefined} from '@feltcoop/felt/util/object.js';
 import type {Partial_Except} from '@feltcoop/felt/util/types.js';
 
 import type {Path_Stats} from './path_data.js';
-import {toPath_Filter} from './path_filter.js';
+import {to_path_filter} from './path_filter.js';
 import type {Path_Filter} from './path_filter.js';
-import {loadGitignoreFilter} from '../utils/gitignore.js';
+import {load_gitignore_filter} from '../utils/gitignore.js';
 
 /*
 
@@ -19,21 +19,21 @@ export interface Watch_Node_Fs {
 	close: () => void;
 }
 
-export interface WatcherChange {
-	type: WatcherChangeType;
+export interface Watcher_Change {
+	type: Watcher_Change_Type;
 	path: string;
 	stats: Path_Stats;
 }
-export type WatcherChangeType = 'create' | 'update' | 'delete';
-export interface WatcherChangeCallback {
-	(change: WatcherChange): void;
+export type Watcher_Change_Type = 'create' | 'update' | 'delete';
+export interface Watcher_Change_Callback {
+	(change: Watcher_Change): void;
 }
 
 export const DEBOUNCE_DEFAULT = 10;
 
 export interface Options {
 	dir: string;
-	on_change: WatcherChangeCallback;
+	on_change: Watcher_Change_Callback;
 	filter: Path_Filter | null | undefined;
 	watch: boolean;
 	debounce: number;
@@ -44,15 +44,15 @@ export const init_options = (opts: Initial_Options): Options => ({
 	watch: true,
 	debounce: DEBOUNCE_DEFAULT,
 	...omit_undefined(opts),
-	filter: opts.filter === undefined ? toDefaultFilter() : opts.filter,
+	filter: opts.filter === undefined ? to_default_filter() : opts.filter,
 });
 
 export const watch_node_fs = (opts: Initial_Options): Watch_Node_Fs => {
 	const {dir, on_change, filter, debounce, watch} = init_options(opts);
-	const watcher = new CheapWatch({dir, filter, watch, debounce});
+	const watcher = new Cheap_Watch({dir, filter, watch, debounce});
 	if (watch) {
-		watcher.on('+', ({path, stats, isNew}) => {
-			on_change({type: isNew ? 'create' : 'update', path, stats});
+		watcher.on('+', ({path, stats, is_new}) => {
+			on_change({type: is_new ? 'create' : 'update', path, stats});
 		});
 		watcher.on('-', ({path, stats}) => {
 			on_change({type: 'delete', path, stats});
@@ -70,4 +70,4 @@ export const watch_node_fs = (opts: Initial_Options): Watch_Node_Fs => {
 	};
 };
 
-const toDefaultFilter = (): Path_Filter => toPath_Filter(loadGitignoreFilter());
+const to_default_filter = (): Path_Filter => to_path_filter(load_gitignore_filter());